requirements.txt
# Dependências principais
ultralytics>=8.1.0
opencv-python>=4.8.0
numpy>=1.22.0
pyyaml>=6.0.0
matplotlib>=3.7.0
pillow>=10.0.0
tqdm>=4.65.0

# Bibliotecas para interface gráfica de anotação
tk; platform_system != "Darwin" or python_version < "3.12"

# PyTorch (as versões específicas serão instaladas conforme o ambiente)
torch>=2.0.0
torchvision>=0.15.0

# Utilidades para processamento de dados
pandas>=2.0.0
scikit-learn>=1.3.0
setuptools~=75.8.0
colorama>=0.4.6

# Biliotecas para o portal de documentações
pygments>=2.19.1
markdown>=3.6
pytest~=8.3.4
requests~=2.32.3
defusedxml~=0.7.1
scipy>=1.13.1
jinja2~=3.1.6
plotly>=5.24.1
dash>=3.0.0
dash_bootstrap_components>=2.0.0